### 2025-08-04 #8702 @editorialbot
[PRE REVIEW]: ProntoPCR: Efficient qPCR Data Analysis Software
**Submitting author:** @MarnieMaddock (<a href="http://orcid.org/0000-0001-6393-4837">Marnie Maddock</a>)
**Repository:** https://github.com/MarnieMaddock/ProntoPCR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
