### 2025-12-21 6f77964
Shield update check from online session
