### 2025-12-21 d62c89e
lazy load server calls
