+++ R/app_function.R
@@ -105,7 +105,7 @@ include_analytics_html <- function() {
 
 #' @export
 ProntoPCR <-  function(...) {
-  #only perform if app is being sed on R
+  #only perform if app is run locally
   if(Sys.getenv("RSCONNECT_SERVER") == ""){
     check_for_updates()
     # Display formal citation message in console
@@ -119,9 +119,6 @@ ProntoPCR <-  function(...) {
             "\n==================================================================\n")
   }
 
-  
-
-  
  #set up UI 
   ui <- fluidPage(
     theme = bslib::bs_theme(version = 4, bootswatch = "pulse"), #theme
@@ -265,51 +262,198 @@ ProntoPCR <-  function(...) {
                        regular = get_font_path("tahoma.ttf"),
                        bold = get_font_path("tahomabd.ttf"))
     
-    # #insert csv file and check that it meets the required formatting
-    csv_data  <- checkCSVfile("file")
-    downloadExampleData("file") #download example data file link
-    # Generate dynamic text input fields for housekeeper genes
-    fileModule <- inputFileServer("file")
+    ## store reactive vals
+    csv_data               <- reactiveVal(NULL)
+    fileModule             <- reactiveVal(NULL)
     
-    # Use the input data module to display inserted file in the UI
-    inputDataServer("inputDataModule", csv_data)
+    wrangled_data_module   <- reactiveVal(NULL)
+    wrangled_data          <- reactiveVal(NULL)
+    filter_condition       <- reactiveVal(NULL)
     
-    #calculate mean housekeepers, delta Cq and fold change dcq
-    wrangled_data_module <- wrangleDataServer("wrangleDataModule", fileModule$save_btn, csv_data$data, fileModule$saved_variables)
+    rep_avg_data           <- reactiveVal(NULL)
     
-    # Display and calculate biological replicate average values for dcq. Save variables
-    wrangled_data <- wrangled_data_module$wrangled_data
-    filter_condition <- wrangled_data_module$filter_condition
+    ddcq_data_module       <- reactiveVal(NULL)
+    average_dcq            <- reactiveVal(NULL)
+    selected_gene          <- reactiveVal(NULL)
+    ddcq_repData           <- reactiveVal(NULL)
     
-    #perform biological replicate calculations
-    DCQ_repData <- repDataServer("rep_data", wrangled_data, filter_condition)
-    #save dcq rep avg data table
-    rep_avg_data <- DCQ_repData$rep_avg_data
+    stats_module           <- reactiveVal(NULL)
     
-    # select groups for ddcq and calculate ddcq for a gene
-    # Display and calculate biological replicate average values for ddcq
-    ddcq_data_module <- ddcqServer("ddcqModule", wrangled_data)
-    average_dcq <- ddcq_data_module$average_dcq
-    selected_gene <- ddcq_data_module$extracted_gene
-    ddcq_rep_module <- DDCQrepServer("ddcqRep", average_dcq, selected_gene)
-    ddcq_repData <- ddcq_rep_module$rep_avg_data_ddcq
+    graph_generated        <- reactiveVal(FALSE)
     
-    #statistics
-    stats <- statsServer("statsModule", values = ddcq_data_module$values, dcq_data = wrangled_data, ddcq_data = average_dcq, ddcq_selected_gene = ddcq_data_module$gene_for_download)
-    selected_stat <- stats$selected_stat
-    stats_gene <- stats$columnInput
-    filter_data_stats <- stats$filter_data_stats
-    group_comparison <- stats$group_comparison
-    comparisonResults <- reactive({
-      stats$comparisonResults()
-    })
-    descriptives_table <- stats$descriptives_table
[…]
