### 2025-12-22 0b0f6ae
Improve shinyapps startup
