+++ R/app_function.R
@@ -118,8 +118,8 @@ ProntoPCR <-  function(...) {
             "\nThank you for supporting open-source research!",
             "\n==================================================================\n")
   }
-
- #set up UI 
+  
+  #set up UI 
   ui <- fluidPage(
     theme = bslib::bs_theme(version = 4, bootswatch = "pulse"), #theme
     include_analytics_html(), # Include Google Analytics
@@ -137,22 +137,22 @@ ProntoPCR <-  function(...) {
         style = "height: 85vh; overflow-y: auto;", # Set the sidebar height and add a scroll bar
         conditionalPanel(condition = "input.tabselected==2 && input.subInput == 2.1",
                          inputFileUI("file")
-                         ), #insert csv file and check that it meets the required formatting, enter housekeeper names and save them
+        ), #insert csv file and check that it meets the required formatting, enter housekeeper names and save them
         #Calculations tab: #Delta Cq tab
         conditionalPanel(condition = "input.tabselected == 3 && input.subPanel == 3.1",
                          wrangleDataSidebar("wrangleDataModule") #display delta Cq data, average housekeepers module
-                        ),
+        ),
         conditionalPanel(condition = "input.tabselected == 3 && input.subPanel == 3.2 && input.subCalc2 == 3",
                          ddcqSidebar("ddcqModule") #display ddcq data module
-                        ),
+        ),
         # Statistics tab
         # Suggested workflow tab
         conditionalPanel(condition = "input.tabselected == 4",
                          statsSidebar("statsModule") #display stats options module
-                        ),
+        ),
         conditionalPanel(condition = "input.tabselected == 5",
                          graphsSidebar("graphsModule") #display graphs options
-                        ),
+        ),
       ), #sidebarPanel closing bracket
       
       mainPanel(
@@ -224,43 +224,55 @@ ProntoPCR <-  function(...) {
   ) #fluidPage close bracket
   
   server <- function(input, output, session) {
-    # Enable automatic font rendering via showtext
-    showtext::showtext_auto()
+
     # Add system fonts
-    sysfonts::font_add("Arial", 
-                       regular = get_font_path("arial.ttf"), 
-                       italic = get_font_path("ariali.ttf"), 
-                       bold = get_font_path("arialbd.ttf"), 
-                       bolditalic = get_font_path("arialbi.ttf"))
-    sysfonts::font_add("Arial Bold", get_font_path("arialbd.ttf"))
-    sysfonts::font_add("Calibri", 
-                       regular = get_font_path("calibri.ttf"), 
-                       italic = get_font_path("calibrii.ttf"), 
-                       bold = get_font_path("calibrib.ttf"), 
-                       bolditalic = get_font_path("calibriz.ttf"))
-    sysfonts::font_add("Times New Roman", 
-                       regular = get_font_path("times.ttf"), 
-                       italic = get_font_path("timesi.ttf"), 
-                       bold = get_font_path("timesbd.ttf"), 
-                       bolditalic = get_font_path("timesbi.ttf"))
-    sysfonts::font_add("Georgia", 
-                       regular = get_font_path("georgia.ttf"), 
-                       italic = get_font_path("georgiai.ttf"), 
-                       bold = get_font_path("georgiab.ttf"), 
-                       bolditalic = get_font_path("georgiaz.ttf"))
-    sysfonts::font_add("Comic Sans MS", 
-                       regular = get_font_path("comic.ttf"), 
[…]
