### 2025-12-22 d990661
add ggtext for markdown
