+++ NAMESPACE
@@ -40,6 +40,7 @@ importFrom(ggplot2,scale_y_continuous)
 importFrom(ggplot2,stat_summary)
 importFrom(ggplot2,theme)
 importFrom(ggpubr,stat_pvalue_manual)
+importFrom(ggtext,element_markdown)
 importFrom(magrittr,"%>%")
 importFrom(rlang,sym)
 importFrom(shinyjs,useShinyjs)
