### 2025-12-22 a31fee4
update refs and affiliations
