+++ paper/paper.bib
@@ -7,7 +7,7 @@
   pages={402-408},
   year={2001},
   issn = {1046-2023},
-  doi = {https://doi.org/10.1006/meth.2001.1262},
+  doi = {10.1006/meth.2001.1262},
   publisher={Elsevier}
 }
 
@@ -17,6 +17,7 @@
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
     year = {2013},
+    doi = {10.32614/R.manuals},
     url = {http://www.R-project.org/},
 }
 
@@ -25,8 +26,14 @@
     author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
     year = {2021},
     note = {R package version 1.6.0},
-    url = {https://CRAN.R-project.org/package=shiny},
+    url = {https://CRAN.R-project.org/package=shiny}
 }
 
 
-add ProntoPCR github and vignette to rference list
+@misc{ProntoPCRHandbook,
+  author = {Maddock, Marnie},
+  title  = {ProntoPCR User Guide and Handbook},
+  year   = {2025},
+  url    = {https://marniemaddock.github.io/ProntoPCR/}
+}
+
