+++ paper/paper.bib
@@ -29,8 +29,7 @@
     url = {https://CRAN.R-project.org/package=shiny}
 }
 
-
-@misc{ProntoPCRHandbook,
+@Manual{ProntoPCRHandbook,
   author = {Maddock, Marnie},
   title  = {ProntoPCR User Guide and Handbook},
   year   = {2025},
