+++ paper/paper.bib
@@ -28,11 +28,3 @@
     note = {R package version 1.6.0},
     url = {https://CRAN.R-project.org/package=shiny}
 }
-
-@Manual{ProntoPCRHandbook,
-  author = {Maddock, Marnie},
-  title  = {ProntoPCR User Guide and Handbook},
-  year   = {2025},
-  url    = {https://marniemaddock.github.io/ProntoPCR/}
-}
-
