+++ tests/testthat/_snaps/windows-4.5/graphs-designAesthetics.md
@@ -1,0 +1,4 @@
+# Graph Design aesthetics
+
+    "Graph_IL6_2025-12-22-13-30-59.svg"
+
