### 2025-12-22 b956cf2
quotes for affils
