### 2025-12-22 cfedbda
update affils
