### 2026-02-07 #8702 @majensen
Hey @georgezakinih - would you be willing to review this interesting Shiny App (#8702) for [JOSS](https://joss.theoj.org)? I can really use your expertise - thanks for considering! --@majensen
