### 2026-02-23 #9949 @fangliu117
This is a useful tool that addresses a genuine need in the qPCR research community. The software is well-structured as an R package, has clear documentation via the Handbook, and provides an accessible GUI for researchers with limited programming experience. The following items should be addressed or clarified:
1. **Installation (macOS)**: Following the README/handbook via the default `devtools::install_github("MarnieMaddock/ProntoPCR")` initially failed on my macOS setup with a download error from GitHub (API tarball). Setting `options(download.file.method = "curl")` before running `devtools::install_github("MarnieMaddock/ProntoPCR", ref = "main", upgrade = "never")` allowed the installation to complete successfully, so suggest to document the specific macOS/restricted-network setups.
2. The GitHub repository currently shows 0 tags and no published releases. A versioned release tag (e.g., v1.0.0) should exist on GitHub so that the reviewed version of the software is clearly identified and permanently referenceable.
3. Add a "**State of the Field**" section or incorporate into the "Statement of Need" to compare ProntoPCR with existing qPCR analysis tools. For instance, relevant tools to consider may include:
- **tidyqpcr** (rOpenSci, R package): tidyverse-based qPCR analysis with MIQE compliance
- **pcr** (CRAN R package): provides [ΔΔCt calculations](https://cran.r-project.org/web/packages/RQdeltaCT/vignettes/my-vignette.html)
- **shinyCurves** ([Olaechea-Lázaro et al., BMC Bioinformatics, 2021](https://pmc.ncbi.nlm.nih.gov/articles/PMC8487674/) Shiny app for multi-source qPCR data
- **FaDA** (Shiny app for general lab data including qPCR)
- **Qupid** (Python/Streamlit ΔΔCt web application)
The reference list will be updated accordingly.
4. Assumptions of the 2^−ΔΔCq method: ProntoPCR implements the Livak & Schmittgen-style 2^−ΔΔCq workflow (Table 1), but the manuscript does not currently document the underlying assumption that amplification efficiencies are close to ideal and similar between target and reference assays. If efficiencies differ, efficiency-corrected approaches (e.g., Pfaffl 2001) are more appropriate. So I recommend adding a brief note about this assumption.
5. Fix the typos and writing issues:
Line 48- Typo: "An example dataset can **de** downloaded" → "can **be** downloaded"
Table 1 caption (line 30)-  Redundant label: "Table 1: Table 1: Description of terms..."
Lines 46~49  The ProntoPCR Handbook is mentioned twice in the same paragraph ("A step-by-step guide on how to install and use ProntoPCR is available: ProntoPCR Handbook" and "A user guide is also available: ProntoPCR Handbook"). These appear to be the same document and should be consolidated.
