+++ paper/paper.bib
@@ -1,14 +1,114 @@
 @article{livak2001analysis,
   title= {Analysis of relative gene expression data using real-time quantitative PCR and the 2- $\Delta$$\Delta$CT method},
-  author={Kenneth J Livak and Thomas D Schmittgen},
+  author={Livak, Kenneth J and Schmittgen, Thomas D},
   journal={Methods},
   volume={25},
   number={4},
   pages={402-408},
   year={2001},
-  issn = {1046-2023},
-  doi = {10.1006/meth.2001.1262},
+  issn={1046-2023},
+  doi={10.1006/meth.2001.1262},
   publisher={Elsevier}
+}
+
+@article{Wallace_2022, 
+ title={tidyqpcr: Quantitative PCR analysis in the tidyverse.}, 
+ author={Wallace, Edward WJ and Haynes, Samuel J}, 
+ journal={Journal of Open Source Software}, 
+ volume={7}, 
+ doi={10.21105/joss.04507}, 
+ number={74}, 
+ publisher={The Open Journal}, 
+ year={2022}, 
+ pages={4507} 
+}
+ 
+@article{Li_2022, 
+  title={qPCRtools: An R package for qPCR data processing and visualization}, 
+  volume={13}, 
+  doi={10.3389/fgene.2022.1002704}, 
+  journal={Frontiers in Genetics}, 
+  publisher={Frontiers Media SA}, 
+  author={Li, Xiang and Wang, Yingmin and Li, Jingyu and Mei, Xinyue and Liu, Yixiang and Huang, Huichuan}, 
+  year={2022} 
+}
+
+@article{Ahmed_2018, 
+ title={pcr: an R package for quality assessment, analysis and testing of qPCR data}, 
+ volume={6},
+ doi={10.7717/peerj.4473}, 
+ journal={PeerJ}, 
+ publisher={PeerJ}, 
+ author={Ahmed, Mahmoud and Kim, Deok Ryong}, 
+ year={2018},
+ pages={e4473} 
+}
+
+@article{Dvinge_2009, 
+  title={HTqPCR: high-throughput analysis and visualization of quantitative real-time PCR data in R}, 
+  volume={25}, 
+  doi={10.1093/bioinformatics/btp578}, 
+  number={24}, 
+  journal={Bioinformatics}, 
+  publisher={Oxford University Press (OUP)}, 
+  author={Dvinge, Heidi and Bertone, Paul}, 
+  year={2009}, 
+  pages={3325–3326} 
+}
+ 
+@article{Ritz_2008, 
+  title={qpcR: an R package for sigmoidal model selection in quantitative real-time polymerase chain reaction analysis}, 
+  volume={24}, 
+  doi={10.1093/bioinformatics/btn227}, 
+  number={13}, 
+  journal={Bioinformatics}, 
+  publisher={Oxford University Press (OUP)}, 
+  author={Ritz, Christian and Spiess, Andrej-Nikolai}, 
+  year={2008}, 
+  pages={1549–1551} 
+}
+
+@article{Olaechea_L_zaro_2021, 
+ title={shinyCurves, a shiny web application to analyse multisource qPCR amplification data: a COVID-19 case study}, 
+ volume={22}, 
+ doi={10.1186/s12859-021-04392-1}, 
+ number={1}, 
+ journal={BMC Bioinformatics}, 
+ publisher={Springer Science and Business Media LLC}, 
+ author={Olaechea-Lázaro, S. and García-Santisteban, I. and Pineda, J. R. and Badiola, I. and Alonso, S. and Bilbao, Jose Ramon and Fernandez-Jimenez, Nora}, 
+ year={2021}
+}
+
+@article{Danger_2021, 
[…]
