+++ paper/paper.md
@@ -36,7 +36,7 @@ Quantitative Polymerase Chain Reaction (qPCR) is an essential technique used to 
 
 Methods of quantifying relative gene expression through qPCR have remained largely unchanged since its implementation in 2001 [@livak2001analysis]. These calculations rely on raw quantification cycle (Cq) results from a qPCR reaction, which can then be normalised in various manners (Table 1). These methods however, are usually performed manually in programs such as Excel. Although this quantification is relatively simple, there are caveats to manual analysis such as being time consuming and prone to human or data entry errors. There is a growing need to automate these tasks, especially for handling large volumes of data. Presenting data in graphical formats is also a laborious task for many samples and targets. Therefore, we developed an application called \texttt{ProntoPCR} to address these issues by seamlessly semi-automating the calculations and analysis of qPCR data. A user can go from raw data to results, including statistics and graphs in just a few clicks.
 
-: Table 1: Description of terms and formulas used in qPCR calculations within ProntoPCR.
+: Description of terms and formulas used in qPCR calculations within ProntoPCR.
 
 |  **Term**            | **ProntoPCR Alias**  | **Formula/Explanation** |
 | :------------------- | :------------- | :---------------------- |
@@ -47,13 +47,20 @@ Methods of quantifying relative gene expression through qPCR have remained large
 | ΔΔCq (Delta Delta Cq) | ddcq_gene | = ΔCq (Gene of Interest) - ΔCq (Control Average) |
 | $2^{−ΔΔCq}$ (Fold Change) | fc_ddcq_gene | = $2^{−ΔΔCq}$ |
 
+# State of the Field
+
+Several tools have been developed to support qPCR analysis across different research contexts and levels of technical expertise. Numerous established R packages support ΔCq and ΔΔCq workflows within a coding environment, including the CRAN/Bioconductor packages *tidyqpcr*, *qPCRtools*, *pcr* and *HTqPCR* [@Wallace_2022; @Li_2022; @Ahmed_2018; @Dvinge_2009]. The *qpcR* package also focuses on amplification curve modelling and efficiency estimation [@Ritz_2008]. While these packages provide flexibility and PCR analysis support, they are primarily designed for scripted workflows and thus require familiarity with R programming. 
+
+Graphical applications with point-and-click interfaces have also been developed. For example, *shinyCurves* provides an interface for amplification curve visualisation and Cq extraction within viral detection contexts [@Olaechea_L_zaro_2021]. *FaDA* is a broader laboratory data analysis platform that supports qPCR, amongst other assay types [@Danger_2021]. Although these applications improve accessibility, they focus on specific tasks rather than offering a single, integrated workflow for routine relative gene expression. *Qupid* is a Python-based application built using Streamlit that provides a local browser-rendered interface for ΔΔCt analysis. It supports flexible data structures through the use of metadata “decorators,” which are text markers inserted within input files to designate assays, normalisers, and replicate groups [@kleinschmidt2023qupid]. In contrast, ProntoPCR is structured around commonly exported instrument formats and guided user selections, reducing the need for manual file modification while integrating normalisation, statistical testing, and graphical output within a single workflow.
+
+
 # Overview
 
[…]
