+++ paper/paper.bib
@@ -90,7 +90,18 @@
   year={2021}, 
   pages={e0261083} 
 }
- 
+
+@article{Rancurel_2019, 
+  title={SATQPCR: Website for statistical analysis of real-time quantitative PCR data}, 
+  volume={46}, 
+  doi={10.1016/j.mcp.2019.07.001}, 
+  journal={Molecular and Cellular Probes}, 
+  publisher={Elsevier BV}, 
+  author={Rancurel, Corinne and van Tran, Trang and Elie, Céline and Hilliou, Frédérique}, 
+  year={2019}, 
+  pages={101418} 
+}
+  
 @software{kleinschmidt2023qupid,
   author  = {Kleinschmidt, Noah},
   title   = {Qupid - Quantitative PCR Interface to Delta-Delta-Ct, a web-application for swift Delta-Delta-Ct analysis of qPCR data.},
