+++ paper/paper.md
@@ -51,7 +51,7 @@ Methods of quantifying relative gene expression through qPCR have remained large
 
 Several tools have been developed to support qPCR analysis across different research contexts and levels of technical expertise. Numerous established R packages support ΔCq and ΔΔCq workflows within a coding environment, including the CRAN/Bioconductor packages *tidyqpcr*, *qPCRtools*, *pcr* and *HTqPCR* [@Wallace_2022; @Li_2022; @Ahmed_2018; @Dvinge_2009]. The *qpcR* package also focuses on amplification curve modelling and efficiency estimation [@Ritz_2008]. While these packages provide flexibility and PCR analysis support, they are primarily designed for scripted workflows and thus require familiarity with R programming. 
 
-Graphical applications with point-and-click interfaces have also been developed. For example, *shinyCurves* provides an interface for amplification curve visualisation and Cq extraction within viral detection contexts [@Olaechea_L_zaro_2021]. The SATqPCR web application further extends this landscape by implementing MIQE-guided analysis, incorporating efficiency correction and geNorm-based reference gene selection within a browser-based interface [@Rancurel_2019]. *FaDA* is a broader laboratory data analysis platform that supports qPCR, amongst other assay types [@Danger_2021]. General online relative gene expression calculators also exist. Although these applications improve accessibility, they focus on specific tasks rather than offering a single, integrated workflow for routine relative gene expression. *Qupid* is a Python-based application built using Streamlit that provides a local browser-rendered interface for ΔΔCt analysis. It supports flexible data structures through the use of metadata “decorators,” which are text markers inserted within input files to designate assays, normalisers, and replicate groups [@kleinschmidt2023qupid]. In contrast, ProntoPCR is structured around commonly exported instrument formats and guided user selections, reducing the need for manual file modification while integrating normalisation, statistical testing, and graphical output within a single workflow.
+Graphical applications with point-and-click interfaces have also been developed. For example, *shinyCurves* provides an interface for amplification curve visualisation and Cq extraction within viral detection contexts [@Olaechea_L_zaro_2021]. The *SATqPCR* web application further extends this landscape by implementing MIQE-guided analysis, incorporating efficiency correction and geNorm-based reference gene selection within a browser-based interface [@Rancurel_2019]. *FaDA* is a broader laboratory data analysis platform that supports qPCR, amongst other assay types [@Danger_2021]. General online relative gene expression calculators also exist. Although these applications improve accessibility, they focus on specific tasks rather than offering a single, integrated workflow for routine relative gene expression. *Qupid* is a Python-based application built using Streamlit that provides a local browser-rendered interface for ΔΔCt analysis. It supports flexible data structures through the use of metadata “decorators,” which are text markers inserted within input files to designate assays, normalisers, and replicate groups [@kleinschmidt2023qupid]. In contrast, ProntoPCR is structured around commonly exported instrument formats and guided user selections, reducing the need for manual file modification while integrating normalisation, statistical testing, and graphical output within a single workflow.
 
 
 # Overview
