### 2026-02-24 76f107b
update docs with installation msg
