### 2026-02-24 cc05175
remove Fig 1 duplicate text
