### 2026-02-24 40a728d
affils update
