### 2026-02-24 #9949 @MarnieMaddock
Thank you @fangliu117 for the review and constructive comments. These will greatly improve the manuscript and implementation of ProntoPCR. In response to your comments:
1. Thank you for flagging this issue. We had not come across this error during our testing. We have added troubleshooting guidance to the README and the handbook to help in this scenario.
2. We have now created a release-candidate tag (v1.0.0-rc1) so that the version under review is clearly identified. The final stable release (v1.0.0) will be created once the review process is complete to ensure it reflects any additional revisions.
3. A State of the Field section has been added to the manuscript. We included comparisons of ProntoPCR to the R packages tidyqpcr, pcr as suggested, and the other relevant packages qPCRtools and HTqPCR. We also mention graphical-based applications, including shinyCurves, FaDA and Quipid, as well as including SATqPCR and general online calculators for reference.
4. We have added an assumptions section to the manuscript acknowledging this important assumption.
5. Typos have been fixed. Thank you for pointing these out.
