### 2026-02-24 #9949 @MarnieMaddock
5. The Table 1, Figure 1 duplicates and redundant phrasing have also been removed.
