### 2026-03-09 #9949 @fangliu117
Thanks @MarnieMaddock for the thorough and responsive revisions. I reviewed the updated manuscript and appreciate that my main requests have been addressed.

The revised manuscript now meets JOSS requirements for a software paper. The State of the Field section properly positions ProntoPCR within the existing landscape, the Assumptions section addresses the key scientific caveat, the reference list is appropriate, and the writing is clean. The software itself functions as described and serves a genuine need in the qPCR research community.

Minor optional polish: consider adjusting ‘Pediatrics department’ → ‘Department of Pediatrics’;
Citation update: The R Core Team citation year is 2013; the convention is to cite the version of R actually used during development.

Hi @majensen I have no further major concerns and the paper fulfills all the criteria of the checklist, please find above updated the review checklist. I recommend acceptance, pending a final tagged stable release and software archiving with an archive DOI (e.g., Zenodo/figshare).
