### 2026-04-15 #9949 @georgezakinih
@majensen and @MarnieMaddock , regarding this guideline in JOSS. Is there a "core library" that the shiny app is built upon?

https://joss.theoj.org/about#ai-policy

"Web-based tools may be considered 'in scope' for JOSS, provided that they meet one or both of the following criteria: 1) they are built around and expose a 'core library' through a web-based experience (e.g., R/[Shiny](https://www.rstudio.com/products/shiny/) applications)"
