### 2026-04-15 #9949 @georgezakinih
@MarnieMaddock, I am trying to deploy the app locally and on a Shiny server (Posit Connect), and I recommend the following:

Dependency Management Recommendation: The application currently relies solely on the DESCRIPTION file for dependency specification, which works well for local installation via devtools::install_github() and shinyapps.io deployment. However, for broader reproducibility and compatibility with enterprise deployment platforms such as Posit Connect, I recommend adding an renv.lock file to the repository. This lockfile pins exact package versions, ensuring consistent behavior across different environments and preventing potential breakage from future CRAN updates.

Generating this file is straightforward—running renv::init() followed by renv::snapshot() in the project directory will capture all dependencies with their current versions. This addition would strengthen the application's long-term maintainability and expand deployment options without affecting the existing installation workflow.

Would you please add the renv.lock to the repo so that I can rebuild and deploy the app?
