+++ paper/paper.bib
@@ -127,7 +127,7 @@
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
-    year = {2013},
+    year = {2025},
     doi = {10.32614/R.manuals},
     url = {http://www.R-project.org/},
 }
