### 2026-04-15 #9949 @MarnieMaddock
@georgezakinih thanks for taking the time to test this on Posit Connect, I really appreciate it, as I don’t have access to that environment myself. I’ve now added an renv.lock file to the repo to pin package versions. Let me know if that doesn't work.
