### 2026-04-15 #9949 @MarnieMaddock
> [@majensen](https://github.com/majensen) and [@MarnieMaddock](https://github.com/MarnieMaddock) , regarding this guideline in JOSS. Is there a "core library" that the shiny app is built upon?
>
> https://joss.theoj.org/about#ai-policy
>
> "Web-based tools may be considered 'in scope' for JOSS, provided that they meet one or both of the following criteria: 1) they are built around and expose a 'core library' through a web-based experience (e.g., R/[Shiny](https://www.rstudio.com/products/shiny/) applications)"

In the case of ProntoPCR, the core library corresponds to the underlying R package code in the R/ directory that implements the qPCR analysis workflow (ΔCq/ΔΔCq processing, replicate handling, statistical testing, and plotting). These are written as modular functions and are not tied to the UI itself. The Shiny app serves as a frontend to expose this functionality, rather than acting as a simple wrapper around existing qPCR processing functions that only do parts of the processing pipeline ProntoPCR can do as a whole. While the app makes use of standard R packages (e.g. dplyr for data wrangling etc), the data processing and analysis workflow itself (e.g. handling of input structure, replicate aggregation, and integration of statistical testing with cutsom visualisation) is implemented within the package. Based on this, my understanding is that the app is built around and exposes this core computational layer, in line with the JOSS scope criteria for web-based tools.
