+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{livak2001analysis,
-  title= {Analysis of relative gene expression data using real-time quantitative PCR and the 2- $\Delta$$\Delta$CT method},
+  title= {Analysis of relative gene expression data using real-time quantitative PCR and the 2^{-\Delta\Delta C_{T}} method},
   author={Livak, Kenneth J and Schmittgen, Thomas D},
   journal={Methods},
   volume={25},
@@ -24,7 +24,7 @@
 }
  
 @article{Li_2022, 
-  title={qPCRtools: An R package for qPCR data processing and visualization}, 
+  title={qPCRtools: An {R} package for qPCR data processing and visualization}, 
   volume={13}, 
   doi={10.3389/fgene.2022.1002704}, 
   journal={Frontiers in Genetics}, 
@@ -34,7 +34,7 @@
 }
 
 @article{Ahmed_2018, 
- title={pcr: an R package for quality assessment, analysis and testing of qPCR data}, 
+ title={{pcr}: an {R} package for quality assessment, analysis and testing of qPCR data}, 
  volume={6},
  doi={10.7717/peerj.4473}, 
  journal={PeerJ}, 
@@ -45,7 +45,7 @@
 }
 
 @article{Dvinge_2009, 
-  title={HTqPCR: high-throughput analysis and visualization of quantitative real-time PCR data in R}, 
+  title={HTqPCR: high-throughput analysis and visualization of quantitative real-time PCR data in {R}}, 
   volume={25}, 
   doi={10.1093/bioinformatics/btp578}, 
   number={24}, 
@@ -57,7 +57,7 @@
 }
  
 @article{Ritz_2008, 
-  title={qpcR: an R package for sigmoidal model selection in quantitative real-time polymerase chain reaction analysis}, 
+  title={qpcR: an {R} package for sigmoidal model selection in quantitative real-time polymerase chain reaction analysis}, 
   volume={24}, 
   doi={10.1093/bioinformatics/btn227}, 
   number={13}, 
