### 2026-06-21 084f6e6
DDCT rendering with latex issue
