+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{livak2001analysis,
-  title= {Analysis of relative gene expression data using real-time quantitative PCR and the 2^{-\Delta\Delta C_{T}} method},
+  title= {Analysis of relative gene expression data using real-time quantitative PCR and the $2^{-\Delta\Delta C_{T}}$ method},
   author={Livak, Kenneth J and Schmittgen, Thomas D},
   journal={Methods},
   volume={25},
