### 2026-06-28 90f642d
Add citation section to README
Added citation section with DOI badge and reference to journal article.
