### 2026-06-28 655837f
Create CITATION.cff for software citation information
Added citation details for the software, including authors, DOI, and preferred citation format.
