+++ CITATION.cff
@@ -1,0 +1,45 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Maddock
+  given-names: Marnie L.
+  orcid: "https://orcid.org/0000-0001-6393-4837"
+- family-names: Dottori
+  given-names: Mirella
+  orcid: "https://orcid.org/0000-0003-0598-4195"
+- family-names: Nettel-Aguirre
+  given-names: Alberto
+  orcid: "https://orcid.org/0000-0001-6604-0652"
+contact:
+- family-names: Maddock
+  given-names: Marnie L.
+  orcid: "https://orcid.org/0000-0001-6393-4837"
+- family-names: Nettel-Aguirre
+  given-names: Alberto
+  orcid: "https://orcid.org/0000-0001-6604-0652"
+doi: 10.5281/zenodo.20789353
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Maddock
+    given-names: Marnie L.
+    orcid: "https://orcid.org/0000-0001-6393-4837"
+  - family-names: Dottori
+    given-names: Mirella
+    orcid: "https://orcid.org/0000-0003-0598-4195"
+  - family-names: Nettel-Aguirre
+    given-names: Alberto
+    orcid: "https://orcid.org/0000-0001-6604-0652"
+  date-published: 2026-06-25
+  doi: 10.21105/joss.09949
+  issn: 2475-9066
+  issue: 122
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 9949
+  title: "ProntoPCR: Efficient qPCR Data Analysis Software"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.09949"
+  volume: 11
+title: "ProntoPCR: Efficient qPCR Data Analysis Software"
