### 2026-06-28 281cae1
Add citation info
