### 2026-06-28 e2af618
update citation info
